Add a output process which can dynamically convert 3 channel pixel data to 4 channel pixel data.

+++ xSchedule/AddReverseDialog.cpp
@@ -29,7 +29,7 @@ END_EVENT_TABLE()
 void AddReverseDialog::ValidateWindow()
 {
     long sc = _outputManager->DecodeStartChannel(TextCtrl_StartChannel->GetValue().ToStdString());
-    StaticText_StartChannel->SetLabel(wxString::Format("%ld", sc));
+    StaticText_StartChannel->SetLabel(wxString::Format("%ld", (long)sc));
     if (sc == 0 || sc > xScheduleFrame::GetScheduleManager()->GetTotalChannels())
     {
         Button_Ok->Enable(false);
