+++ xSchedule/SetDialog.cpp
@@ -102,7 +102,7 @@ SetDialog::SetDialog(wxWindow* parent, OutputManager* outputManager, std::string
 void SetDialog::ValidateWindow()
 {
     long sc = _outputManager->DecodeStartChannel(TextCtrl_StartChannel->GetValue().ToStdString());
-    StaticText_StartChannel->SetLabel(wxString::Format("%ld", sc));
+    StaticText_StartChannel->SetLabel(wxString::Format("%ld", (long)sc));
     if (sc == 0 || sc > xScheduleFrame::GetScheduleManager()->GetTotalChannels())
     {
         Button_Ok->Enable(false);
